Enhance the logic to check capability of bulk counter poll (#39)
Enhance the logic to check the capability of the bulk counter poll in `bulkAddObject`
1. Check bulk capability using m_vendorSai->queryStatsCapability before calling SAI API to check capability
2. Check single call capability before adding it to `m_objectIdsMap`
3. Reduce the log severity to `INFO` when a counter is not supported to be bulk-polled to avoid too many log messages
4. Remove redundant code for buffer pool counters

Details:
In `bulkAddObject` it enables counter polling for a set of objects. A brief flow is the following:
1. Check the object type's counter capability of single operation (`SAI_STATS_MODE_READ_AND_CLEAR` or `SAI_STATS_MODE_READ`) using `updateSupportedCounters` (which eventually calls `query_stats_capability`). All counters that do not support the capability will be removed.
2. Check the object type's counter capability of the corresponding bulk operation (`SAI_STATS_MODE_BULK_READ_AND_CLEAR` and `SAI_STATS_MODE_BULK_READ` respectively).
 Add the counters to `m_objectIdsMap` if they do not support bulk call.
3. Check the bulk operation capability on the group of objects (since some vendors support bulk operation only on a part of objects). Add the objects to bulk context if all objects support bulk operation. The objects will be polled using bulk operation.
4. Check the single call capability on each object (since some vendors support counters only on a part of objects). The objects that do not support single call will not be polled at all.

Possible error messages in vendor SAI:
An error message can be observed on some platforms if an unsupported counter-polling SAI API is called.
This could happen if the vendor SAI supports a counter polling operation only on part of objects.
In this case, the function `query_stats_capability` returns that it supports the operation on the object type. FlexCounter needs to call the counter polling SAI API to check whether the operation is supported on each object. An error message could be seen if vendor SAI prints it.