Update upstream pipeline name to fix VS test issue. (#1638)
The VS test stage of PR tests pipeline has been failing. The reason is that one of the VS test step is to download ubuntu deb packages built from an upstream pipeline. The upstream pipeline has been renamed from sonic-net.sonic-buildimage-ubuntu20.04 to sonic-net.sonic-buildimage-ubuntu22.04 because of the deprecation of Ubuntu 20.04.

This pipeline needs to be updated according to use the new upstream pipeline name for artifacts downloading.

This change also deprecated a parameter `sonic_buildimage_ubuntu20_04 which is not really necessary.

 * Fixed conflicts in .azure-pipelines/test-docker-sonic-vs-template.yml

Signed-off-by: Xin Wang <xiwang5@microsoft.com>
Signed-off-by: Prabhat Aravind <paravind@microsoft.com>
Co-authored-by: Xin Wang <xiwang5@microsoft.com>